python_backend/pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.5",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "uvicorn[standard]>=0.40.0",
]
